requirements.txt
# Core Dependencies
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.6
langchain-community==0.0.20

# LLM Integration
ollama==0.1.6

# Voice Processing
faster-whisper==0.10.0
sounddevice==0.4.6
scipy==1.12.0
numpy==1.26.3

# UI Framework
gradio==4.16.0

# Google Gemini (Optional)
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3